requirements.txt
anthropic
ultralytics
fastapi
uvicorn
python-dotenv
sgp4
requests
numpy